Re-factoring and re-arrange the code for norm stds
