+++ includes/shell.h
@@ -6,7 +6,7 @@
 /*   By: tsomacha <tsomacha@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/04 05:39:43 by tsomacha          #+#    #+#             */
-/*   Updated: 2025/05/09 04:16:47 by tsomacha         ###   ########.fr       */
+/*   Updated: 2025/05/09 04:37:46 by tsomacha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -210,6 +210,16 @@ void	perror_exit(const char *msg);
 void	stat_error(char *path, char c);
 void	p_exe_error(char *command, int err, t_initenv *initenv);
 
+/**
+ * Clenaer funtions
+*/
+int		clear_and_exit(t_shell *mini);
+int		clear_commands(t_cmd *cmds);
+int		clear_tokens(t_list *tokens);
+int		clear_array(char **array);
+void	clear_env(t_env *env);
+void	free_redirections(t_redir *redir_list);
+
 char	*extract_env_value(t_initenv *initenv, char *name);
 char	**copy_env(t_env *env);
 int		activate_shell(char *input, t_initenv *env);
@@ -217,15 +227,12 @@ int		builtin_cd(t_shell *mini);
 int		builtin_echo(t_cmd *cmd);
 int		builtin_env(t_shell *mini);
 int		check_builtin(t_cmd *current, t_shell *mini);
-int		clear_and_exit(t_shell *mini);
-int		clear_array(char **array);
 int		execute(t_shell *mini);
 int		exit_mini(t_shell *mini);
 int		ft_isquoted(const char *str, int n);
 t_cmd	*list_add_command(t_cmd *cmds, t_cmd *node);
 void	add_to_list(t_env **env, char *content);
 void	builtin_unset(t_shell *mini);
-void	clear_env(t_env *env);
 void	list_env(t_env **env, char **envp);
 
 #endif
